requirements.txt
atlasclient
mysql-connector-python